+++ src/exosNonAmcNonInteractifs.md
@@ -34,11 +34,11 @@
 |6I11|5G31|4G40-2|3L10|2N43-7|||betaModèle40_tableau_proportionnalite|
 |6M11-2|5G40-1|4G40-MG32|3L13-1|2N52-1|||betaModèle41_tableau_signes_variations|
 |6M22-1|5G40-2|4G41|3L13-2|2N52-2|||betaProbaAouB|
-|6M22-2|5G40|4G51|3L13|2N52-4|||betaProbabilites|
-|6M25|5G41|4G52|3L14-1|2N60-4|||betaPuissances|
-|6M31-2|5G42|4I1|3L14|2N61-2|||betarotation3d|
-|6N10-2|5G51|4L10-1|3L15|2N61-4|||betaSpline|
-|6N10-3|5L10-1|4L10|3S20||||betaSys2x2CombLin|
+|6M22-2|5G40|4G51|3L14-1|2N52-4|||betaProbabilites|
+|6M25|5G41|4G52|3L14|2N60-4|||betaPuissances|
+|6M31-2|5G42|4I1|3L15|2N61-2|||betarotation3d|
+|6N10-2|5G51|4L10-1|3S20|2N61-4|||betaSpline|
+|6N10-3|5L10-1|4L10|||||betaSys2x2CombLin|
 |6N10-4|5L10-3|4L11|||||betaTracerParabole|
 |6N10-5|5L10-4|4L13-0|||||betatrinome|
 |6N10|5L12-1|4L13-1|||||moule_a_exo_mathalea|
@@ -48,12 +48,12 @@
 |6N14|5L14-4|4L15-0|||||c3C11|
 |6N20-1|5L14-5|4L15-1|||||c3N10|
 |6N22-1|5L14-6|4L16|||||c3N20|
-|6N22-2|5L15|4L20|||||c3N22|
-|6N23-0|5M10|4P10-1|||||c3N23|
-|6N23-1|5N11-1|4P10-2|||||CM020|
-|6N23-2|5N11-2|4P10|||||CM021|
-|6N23-3|5N11-3|4S10|||||PEA11-1|
-|6N23-4|5N11-4|4S11|||||PEA11|
+|6N22-2|5L15|4P10-1|||||c3N22|
+|6N23-0|5M10|4P10-2|||||c3N23|
+|6N23-1|5N11-1|4P10|||||CM020|
+|6N23-2|5N11-2|4S10|||||CM021|
+|6N23-3|5N11-3|4S11|||||PEA11-1|
+|6N23-4|5N11-4||||||PEA11|
 |6N23|5N14-2||||||P003|
 |6N24-1|5N14||||||P004|
 |6N31-1|5N15||||||P005|
